higherOrder HW all tests passing
