+++ 06week/higherOrder/package.json
@@ -1,0 +1,9 @@
+{
+  "scripts": {
+    "start": "node higherOrder",
+    "test": "mocha"
+  },
+  "devDependencies": {
+    "mocha": "^5.2.0"
+  }
+}
